* Changing copy for syncrhonize (#81)

* Adding error fail message (#79)

* Grouping execution - turning red errors (#61)

* Grouping

* Quick fix to translate / into - (#85)

* Add support for variable env vars (#84)

* Adding option to include GHVars

* Adding validation for repo file, fixing readme

* Update README.md

* Update README.md

* Adjusted for snippet consistency

+++ operations/_scripts/generate/generate_app_repo.sh
@@ -17,6 +17,9 @@ fi
 
 cp -rf "$TARGET_PATH"/* "${GITHUB_ACTION_PATH}/operations/deployment/ansible/app/${GITHUB_REPO_NAME}/"
 
-echo "Copying checked in env file from repo to Ansible deployment path"
-
-cp "$TARGET_PATH/$REPO_ENV" "${GITHUB_ACTION_PATH}/operations/deployment/ansible/repo.env"
+if [ -s "$TARGET_PATH/$REPO_ENV" ]; then
+  echo "Copying checked in env file from repo to Ansible deployment path"
+  cp "$TARGET_PATH/$REPO_ENV" "${GITHUB_ACTION_PATH}/operations/deployment/ansible/repo.env"
+else
+  echo "Checked in env file from repo is empty or couldn't be found"
+fi
